requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-identity>=1.12.0
azure-keyvault-secrets>=4.7.0
pyairtable==3.0.1
python-dotenv==1.0.1
googlemaps==4.10.0
azure-ai-documentintelligence==1.0.0
